+++ OUDS/Core/Components/Sources/Controls/Checkbox/OUDSCheckbox.swift
@@ -63,11 +63,11 @@ import SwiftUI
 ///
 /// ## Suggestions
 ///
-/// According to the [documentation](https://unified-design-system.orange.com/472794e18/p/09d860-checkbox/t/14bf4bd854), the checkbox by default must be used in unselected state.
+/// According to the [documentation](https://r.orange.fr/r/S-ouds-doc-checkbox), the checkbox by default must be used in unselected state.
 ///
 /// ## Design documentation
 ///
-/// [unified-design-system.orange.com](https://unified-design-system.orange.com/472794e18/p/23f1c1-checkbox)
+/// [unified-design-system.orange.com](https://r.orange.fr/r/S-ouds-doc-checkbox)
 ///
 /// ## Themes rendering
 ///