+++ OUDS/Core/Components/Sources/Indicators/Tag/OUDSTag.swift
@@ -117,7 +117,7 @@ import SwiftUI
 ///
 /// ## Design documentation
 ///
-/// [unified-design-system.orange.com](https://unified-design-system.orange.com/472794e18/p/7565ce-tag)
+/// [unified-design-system.orange.com](https://r.orange.fr/r/S-ouds-doc-tag)
 ///
 /// ## Themes rendering
 ///