+++ OUDS/Core/Components/Sources/Layouts/Divider/OUDSVerticalDivider.swift
@@ -27,7 +27,7 @@ import SwiftUI
 ///
 /// ## Design documentation
 ///
-/// [unified-design-system.orange.com](https://unified-design-system.orange.com/472794e18/p/629e1b-divider)
+/// [unified-design-system.orange.com](https://r.orange.fr/r/S-ouds-doc-divider)
 ///
 /// ## Themes rendering
 ///