+++ OUDS/Core/Components/Sources/Navigations/Link/OUDSLink.swift
@@ -47,7 +47,7 @@ import SwiftUI
 ///
 /// ## Design documentation
 ///
-/// [unified-design-system.orange.com](https://unified-design-system.orange.com/472794e18/p/31c33b-link)
+/// [unified-design-system.orange.com](https://r.orange.fr/r/S-ouds-doc-link)
 ///
 /// ## Themes rendering
 ///